refactor(dsg): use functions map instead of switch case
