+++ packages/amplication-data-service-generator/src/server/resource/dto/create-enum-dto.ts
@@ -2,7 +2,7 @@ import { namedTypes, builders } from "ast-types";
 import { pascalCase } from "pascal-case";
 import { types, EntityField, Entity } from "@amplication/code-gen-types";
 
-import { createEnumName } from "../../prisma/create-prisma-schema";
+import { createEnumName } from "../../prisma/create-prisma-schema-fields";
 
 export type StringLiteralEnumMember = namedTypes.TSEnumMember & {
   initializer: namedTypes.StringLiteral;
