+++ packages/amplication-data-service-generator/src/server/resource/dto/graphql-field-decorator/create-graphql-field-decorator.ts
@@ -5,7 +5,7 @@ import {
   ScalarField,
   ScalarType,
 } from "prisma-schema-dsl-types";
-import { createEnumName } from "../../../prisma/create-prisma-schema";
+import { createEnumName } from "../../../prisma/create-prisma-schema-fields";
 import { Entity, EntityField } from "@amplication/code-gen-types";
 import { isRelationField, isToManyRelationField } from "../../../../util/field";
 import {
