+++ packages/amplication-data-service-generator/src/server/seed/create-seed.spec.ts
@@ -14,7 +14,7 @@ import {
 } from "./create-seed";
 import { DEFAULT_USER_ENTITY } from "../user-entity";
 import { createEnumMemberName } from "../resource/dto/create-enum-dto";
-import { createEnumName } from "../prisma/create-prisma-schema";
+import { createEnumName } from "../prisma/create-prisma-schema-fields";
 import { memberExpression } from "../../util/ast";
 
 const EXAMPLE_ENTITY_FIELD_NAME = "exampleEntityField";
